move unit tests for proxy invocation event to apigateway tests
