+++ tests/unit/test_apigateway.py
@@ -16,6 +16,7 @@ from localstack.services.apigateway.helpers import (
     extract_query_string_params,
     get_resource_for_path,
 )
+from localstack.services.apigateway.integration import LambdaProxyIntegration
 from localstack.services.apigateway.invocations import (
     ApiInvocationContext,
     RequestValidator,
@@ -426,3 +427,92 @@ def test_create_invocation_headers():
     }
     headers = create_invocation_headers(invocation_context)
     assert headers == {"X-Header": "foobar", "X-Custom": "'Event'"}
+
+
+class TestApigatewayEvents:
+    def test_construct_invocation_event(self):
+        tt = [
+            {
+                "method": "GET",
+                "path": "http://localhost.localstack.cloud",
+                "headers": {},
+                "data": None,
+                "query_string_params": None,
+                "is_base64_encoded": False,
+                "expected": {
+                    "path": "http://localhost.localstack.cloud",
+                    "headers": {},
+                    "multiValueHeaders": {},
+                    "body": None,
+                    "isBase64Encoded": False,
+                    "httpMethod": "GET",
+                    "queryStringParameters": None,
+                    "multiValueQueryStringParameters": None,
+                },
+            },
+            {
+                "method": "GET",
+                "path": "http://localhost.localstack.cloud",
+                "headers": {},
+                "data": None,
+                "query_string_params": {},
+                "is_base64_encoded": False,
+                "expected": {
+                    "path": "http://localhost.localstack.cloud",
+                    "headers": {},
+                    "multiValueHeaders": {},
+                    "body": None,
+                    "isBase64Encoded": False,
+                    "httpMethod": "GET",
+                    "queryStringParameters": None,
+                    "multiValueQueryStringParameters": None,
+                },
+            },
+            {
+                "method": "GET",
+                "path": "http://localhost.localstack.cloud",
+                "headers": {},
+                "data": None,
+                "query_string_params": {"foo": "bar"},
+                "is_base64_encoded": False,
+                "expected": {
+                    "path": "http://localhost.localstack.cloud",
+                    "headers": {},
+                    "multiValueHeaders": {},
+                    "body": None,
+                    "isBase64Encoded": False,
+                    "httpMethod": "GET",
+                    "queryStringParameters": {"foo": "bar"},
+                    "multiValueQueryStringParameters": {"foo": ("bar",)},
+                },
+            },
+            {
+                "method": "GET",
+                "path": "http://localhost.localstack.cloud?baz=qux",
+                "headers": {},
+                "data": None,
+                "query_string_params": {"foo": "bar"},
+                "is_base64_encoded": False,
+                "expected": {
+                    "path": "http://localhost.localstack.cloud?baz=qux",
+                    "headers": {},
+                    "multiValueHeaders": {},
+                    "body": None,
+                    "isBase64Encoded": False,
+                    "httpMethod": "GET",
+                    "queryStringParameters": {"foo": "bar"},
+                    "multiValueQueryStringParameters": {"foo": ("bar",)},
+                },
+            },
+        ]
+
+        for t in tt:
+            result = LambdaProxyIntegration.construct_invocation_event(
+                t["method"],
+                t["path"],
+                t["headers"],
+                t["data"],
+                t["query_string_params"],
+                t["is_base64_encoded"],
+            )
+            assert result == t["expected"]
